requirements.txt
adium-theme-ubuntu==0.3.4
certifi==2017.4.17
click==6.7
Flask==0.12.1
future==0.16.0
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
pygraphviz==1.3.1
python-telegram-bot==6.0.3
six==1.10.0
transitions==0.5.0
unity-lens-photos==1.0
Werkzeug==0.12.2
